also move slakh fixtures
