+++ tests/data/conftest.py
@@ -9,6 +9,14 @@ IKALA_TEST_INDEX = json.load(open(RESOURCES_PATH / "data" / "ikala" / "dummy_ind
 MAESTRO_TEST_INDEX = json.load(open(RESOURCES_PATH / "data" / "maestro" / "dummy_index.json"))
 METADATA_TEST_INDEX = json.load(open(RESOURCES_PATH / "data" / "maestro" / "maestro-v2.0.0.json"))
 MEDLEYDB_PITCH_TEST_INDEX = json.load(open(RESOURCES_PATH / "data" / "medleydb_pitch" / "dummy_index.json"))
+SLAKH_TEST_INDEX = json.load(open(RESOURCES_PATH / "data" / "slakh" / "dummy_index.json"))
+
+
+@pytest.fixture  # type: ignore[misc]
+def mock_slakh_index() -> None:  # type: ignore[misc]
+    with mock.patch("mirdata.datasets.slakh.Dataset.download"):
+        with mock.patch("mirdata.datasets.slakh.Dataset._index", new=SLAKH_TEST_INDEX):
+            yield
 
 
 @pytest.fixture  # type: ignore[misc]
